requirements.txt
firecrawl-py>=4.10.3
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0.0
requests>=2.31.0
playwright>=1.40.0
pandas>=2.0.0
pdfminer.six>=20221105
